agentdemo/requirements.txt
# agentdemo 依赖包
# 多 Agent 智能系统 - 增强版

# 核心依赖
openai>=1.0.0
asyncio
aiohttp>=3.8.0

# 数据处理
pydantic>=2.0.0
pyyaml>=6.0
ujson>=5.0.0

# 日志和监控
structlog>=23.0.0
prometheus-client>=0.17.0

# 工具库
click>=8.0.0
rich>=13.0.0
tqdm>=4.65.0

# 测试和开发
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0

# 类型提示
typing-extensions>=4.0.0

# 其他工具
python-dotenv>=1.0.0
pathlib2>=2.3.0; python_version < '3.4'
